requirements.txt
Django==1.5.5
South==0.8.2
Unidecode==0.04.14
django-dajax==0.9.2
django-dajaxice==0.5.5
django-debug-toolbar==1.0.1
django-disqus==0.4.1
django-haystack==2.1.0
django-reversion==1.8.0
django-rosetta==0.7.3
django-smart-selects==1.0.9
django-taggit==0.11.2
django-tinymce==1.5.2
psycopg2==2.5.2
pysolr==3.2.0
sorl-thumbnail==11.12

